+++ tests/test_generators.py
@@ -1888,56 +1888,6 @@ metadata={'sa': Column(ForeignKey('simple_containers.id'))})
 metadata={'sa': relationship('SimpleContainers', back_populates='simple_items')})
             """)
 
-    def test_onetomany_mandatory(self, generator: CodeGenerator) -> None:
-        Table(
-            'simple_items', generator.metadata,
-            Column('id', INTEGER, primary_key=True),
-            Column('container_id', INTEGER, nullable=False),
-            ForeignKeyConstraint(['container_id'], ['simple_containers.id']),
-        )
-        Table(
-            'simple_containers', generator.metadata,
-            Column('id', INTEGER, primary_key=True)
-        )
-
-        validate_code(generator.generate(), """\
-            from __future__ import annotations
-
-            from dataclasses import dataclass, field
-            from typing import List
-
-            from sqlalchemy import Column, ForeignKey, Integer
-            from sqlalchemy.orm import registry, relationship
-
-            mapper_registry = registry()
-
-
-            @mapper_registry.mapped
-            @dataclass
-            class SimpleContainers:
-                __tablename__ = 'simple_containers'
-                __sa_dataclass_metadata_key__ = 'sa'
-
-                id: int = field(init=False, metadata={'sa': Column(Integer, primary_key=True)})
-
-                simple_items: List[SimpleItems] = field(default_factory=list, \
-metadata={'sa': relationship('SimpleItems', back_populates='container')})
-
-
-            @mapper_registry.mapped
-            @dataclass
-            class SimpleItems:
-                __tablename__ = 'simple_items'
-                __sa_dataclass_metadata_key__ = 'sa'
-
-                id: int = field(init=False, metadata={'sa': Column(Integer, primary_key=True)})
-                container_id: int = field(\
-metadata={'sa': Column(ForeignKey('simple_containers.id'), nullable=False)})
-
-                container: SimpleContainers = field(\
-metadata={'sa': relationship('SimpleContainers', back_populates='simple_items')})
-            """)
-
     def test_manytomany(self, generator: CodeGenerator) -> None:
         Table(
             'simple_items', generator.metadata,